Move non-pure file stuff to a seperate file.
