+++ src/Paths/Directory.ts
@@ -1,4 +1,3 @@
-import { allContentFiles } from "./ContentFile";
 import type { BaseServer } from "../Server/BaseServer";
 import type { FilePath } from "./FilePath";
 import { escapeRegExp } from "lodash";
@@ -95,20 +94,6 @@ export function getFirstDirectoryInPath(path: FilePath | Directory): Directory |
   const firstSlashIndex = path.indexOf("/");
   if (firstSlashIndex === -1) return null;
   return path.substring(0, firstSlashIndex + 1) as Directory;
-}
-
-export function getAllDirectories(server: BaseServer): Set<Directory> {
-  const dirSet = new Set([root]);
-  function peel(path: FilePath | Directory) {
-    const lastSlashIndex = path.lastIndexOf("/", path.length - 2);
-    if (lastSlashIndex === -1) return;
-    const newDir = path.substring(0, lastSlashIndex + 1) as Directory;
-    if (dirSet.has(newDir)) return;
-    dirSet.add(newDir);
-    peel(newDir);
-  }
-  for (const [filename] of allContentFiles(server)) peel(filename);
-  return dirSet;
 }
 
 // This is to validate the assertion earlier that root is in fact a Directory
